startasync

+++ bd.py
@@ -1,4 +1,5 @@
 import sqlite3 as sl
+# import timemod as tm
 
 def new():
     tskdb = sl.connect('tasksdatabase.db')
@@ -9,8 +10,8 @@ def new():
         month INT,
         day INT,
         hour INT,
-        minute INT
-        user_id INT,
+        minute INT,
+        user_id INT
     )""")
     tskdb.commit()
 
@@ -28,21 +29,27 @@ def printtasks(ID):
     tskdb = sl.connect('tasksdatabase.db')
     cur = tskdb.cursor()
     alltasks = []
-    for value in cur.execute("SELECT * FROM tasks ORDER BY year, month, day, hour, minute"):
+    for value in cur.execute(f"SELECT * FROM tasks WHERE user_id ='{ID}' ORDER BY year, month, day, hour, minute"):
         alltasks.append(value)
     return alltasks
 
 def remove(TSK, ID):
     tskdb = sl.connect('tasksdatabase.db')
     cur = tskdb.cursor()
-    cur.execute(f"DELETE FROM tasks WHERE task = '{str(TSK)}'")
+    cur.execute(f"DELETE FROM tasks WHERE task = '{str(TSK)}' AND user_id ='{ID}'")
     tskdb.commit()
 
-def uptask(ID):
+def uptask():
     tskdb = sl.connect('tasksdatabase.db')
     cur = tskdb.cursor()
-    cur.execute("SELECT * FROM tasks ORDER BY year, month, day, hour, minute")
+    cur.execute(f"SELECT * FROM tasks ORDER BY year, month, day, hour, minute")
     return(cur.fetchone())
+
+
+
+
+
+
 
 # if cur.fetchone() is None:
 #     cur.execute(f"INSERT INTO tasks VALUES (?, ?, ?, ?, ?, ?)", (tsk, h, mo, d, h, mi))
